[step2] refactor : 상태를 Closed와 Open으로 분류하여 중복메서드 제거, 패키지 분리
