+++ src/main/java/domain/state/closed/ClosedFrameException.java
@@ -1,4 +1,4 @@
-package domain.state;
+package domain.state.closed;
 
 public class ClosedFrameException extends RuntimeException {
