+++ src/main/java/domain/state/closed/Strike.java
@@ -1,0 +1,10 @@
+package domain.state.closed;
+
+public class Strike extends Closed {
+    private static final String STRIKE_SYMBOL = "X";
+
+    @Override
+    public String printState() {
+        return STRIKE_SYMBOL;
+    }
+}
