+++ src/main/java/domain/state/open/Open.java
@@ -1,0 +1,11 @@
+package domain.state.open;
+
+import domain.state.State;
+
+abstract class Open implements State {
+
+    @Override
+    public boolean isClosed() {
+        return Boolean.FALSE;
+    }
+}
