+++ src/test/java/domain/state/closed/StrikeTest.java
@@ -1,6 +1,7 @@
-package domain.state;
+package domain.state.closed;
 
 import domain.Pins;
+import domain.state.State;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
